Some setup for the basic game
At this point, my worry is that I need to rebalance the cards
substantially enough that the tutorial may require major tweaks.

For example, I feel like the Water Drop card is pretty good, on the
one hand... but then on the other hand, the Netting Shield is even
better, maybe. Its cost might have to be increased to 2, and the Water
Drop maybe completely excised from the starting deck? Maybe it should
be a "somewhat valuable" card?


Former-commit-id: 3f56683b0bdc2d5d00fb6bd479dbb033133f7d1e